+++ lib/core/domain/ambee_api/repository/pollen_repository.dart
@@ -2,7 +2,7 @@ import '../models/pollen_model.dart';
 
 abstract class PollenRepository {
   Future<PollenModel> fetchData({
-    required int latitude,
-    required int longitude,
+    required double latitude,
+    required double longitude,
   });
 }
